refactor(sip-message): ADR-0025 M12 teardown — delete legacy, privatize, MessageCore
The typed value is now the only currency: `SipRequest` / `SipResponse` are a
start line over a private `MessageCore`, the mandatory headers are stored once
as typed values, and every legacy access path is gone.

Deleted: the `message_helpers` namespace (`get_header`/`get_headers`/
`set_header`/`remove_header`, `extract_tag`/`strip_tag`/`extract_contact_uri`/
`extract_name_addr_uri`, `parse_sip_uri`/`parse_uri_params`/`extract_host_port`/
`ParsedSipUri`, `parse_via_params`/`ViaParams`/`via_sent_by`/
`stamp_received_rport_on_via`, `first_route_is_loose`/
`strip_route_uri_to_request_uri`, `header_param_value`, `name_matches`);
`serialize_request_parts` / `serialize_response_parts`; the old
`types::{NameAddr, Via, Contact, CSeq, RequestUri, Uri, Rack, Replaces, ReferTo,
Params, ParamValue}`; `TypedHeader` + `SipMessage::typed`/`get_header`/
`has_header`; `structured_headers::{parse_rack, parse_replaces, parse_refer_to,
parse_param_list}`; and `template::HeaderClass`, which was the same axis as
`header::HeaderClass` under two names.

Promoted out of the deleted namespace, each its own concern:
`sip_message::{emergency, param_codec, preparse, reject_503}`.

The parse populates `CoreHeaders` directly — no second field set to keep in
step — and the scan-side parameter list is the ordered small-vec, so a value
read off a message keeps its wire order and spelling. The datagram is
`msg.image()`; `msg.raw(HeaderName)` stays the header escape hatch.

Acceptance: `grep -rE 'get_header\(|message_helpers::' crates --include='*.rs'`
returns nothing. Workspace: 2105 tests passed, 0 failed. Allocations per
message are down in every budgeted case (decode 14→12, proxy hop 27→22,
build 48→45).

+++ crates/loadgen/tests/smoke.rs
@@ -1740,12 +1740,12 @@ async fn loadgen_inprocess_endurance_lossy() {
     let failed = reporter.total_calls().saturating_sub(total);
     let leaked = b2bua.metrics().creations_total().saturating_sub(b2bua.metrics().removals_total());
     assert!(
-        leaked <= failed as u64,
+        leaked <= failed,
         "SUT leaked {leaked} calls but only {failed} failed under loss — a SUCCESSFUL call leaked:\n{}",
         reporter.render_prometheus()
     );
     assert!(
-        b2bua.active_calls() as u64 <= failed as u64,
+        b2bua.active_calls() as u64 <= failed,
         "SUT holds {} live calls vs {failed} failed",
         b2bua.active_calls()
     );
